PowerBI/Apache Airflow/requirements.txt
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
joblib>=1.3.0
apache-airflow>=2.7.0
apache-airflow-providers-postgres>=5.6.0
psycopg2-binary>=2.9.0
docker>=6.1.0

# Fish Weight Prediction API Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
requests>=2.31.0
